Move client/platform out of traits directory
