--- components/trussed/src/traits.rs
@@ -1,3 +1,0 @@
-
-pub mod client;
-pub mod platform;
