+++ platform/lpc55/src/solo_trussed.rs
@@ -4,7 +4,7 @@ use crate::hal::{
 };
 use board_traits::buttons::{Press, Edge};
 use board_traits::rgb_led::RgbLed;
-use trussed::traits::platform::{
+use trussed::platform::{
     ui,
     reboot,
     consent,
@@ -62,7 +62,7 @@ RGB: RgbLed,
     }
 }
 
-impl<BUTTONS, RGB> trussed::traits::platform::UserInterface for UserInterface<BUTTONS,RGB>
+impl<BUTTONS, RGB> trussed::platform::UserInterface for UserInterface<BUTTONS,RGB>
 where
 BUTTONS: Press + Edge,
 RGB: RgbLed,
